✨ feat: refactor invoice and quote services to use shared formatPattern and getInvertColor functions
